feat: add dynamic route to see artist details dynamically by clicking their card on all artist page
