--- app/(marketing)/(all)/all-artists/loading.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/components/Loading";
