+++ app/(marketing)/(all)/artists/[slug]/_components/ArtistBiography.tsx
