+++ app/(marketing)/(all)/artists/[slug]/_components/Discography.tsx
