+++ app/(marketing)/(all)/artists/[slug]/_components/Merchandise.tsx
