+++ app/(marketing)/(all)/artists/[slug]/_components/Stats.tsx
