+++ app/(marketing)/(all)/artists/page.tsx
@@ -1,25 +1,23 @@
-import Newsletter from "@/components/Newsletter";
-import ArtistBiography from "./_components/ArtistBiography";
-import Discography from "./_components/Discography";
-import Intro from "./_components/Intro";
-import ArtistMerchandise from "./_components/Merchandise";
-import Service from "./_components/Service";
-import Stats from "./_components/Stats";
-import Studio from "./_components/Studio";
-import Testimonial from "./_components/Testimonial";
+import { Heading } from "@/components/Heading";
+import Intro from "../artists/_components/Intro";
+import Container from "@/components/layout/Container";  
+import Footer from "@/components/layout/Footer";
+import ArtistListSection from "../artists/_components/ArtistListSection";
 
 export default function ArtistsPage() {
-  return (
-    <>
-      <Intro />
-      <Stats />
-      <Service />
-      <ArtistBiography />
-      <Discography />
-      <Studio />
-      <Testimonial />
-      <ArtistMerchandise />
-      <Newsletter />
-    </>
-  );
+    return (
+        <>
+            <Intro />
+            <Container>
+                <div className="my-24 max-w-7xl mx-auto">
+                    <Heading>
+                        All Artists
+                    </Heading>
+                </div>
+                <ArtistListSection/>
+            </Container>
+            <Footer/>
+        </>
+    );
 }
+
